Retain last focussed control when navigating back
Navigating back: focus is returned because the ViewModel on top of stack closes, or UserInterface focus is returned (a dialog closes).

Retaining last focus also works for ItemControls through specific code. Requires that a button is in the DataTemplate.
Efforts to make the focusing of last control work in conjunction with NavigatorViewModels and their current ViewModel.

Two possible arrangements using NavigatorViewModels:

1. The View associated with the NavigatorViewModel inherits from UserControl, the Views that are inside it inherit from View.
	- In this case, the NavigatorViewModel must not be focusable.
	- Examples: MainView, CombatView

2. The View associated with the NavigatorViewModel inherits from View, the Views that are inside it inherit from NestedView.
	- The NavigatorViewModel contains controls that can be focussed.
	- The navigation cannot actually use the stack, as the NestedViews themselves do not keep track of the last focussed control.
	- Only navigation using replace is allowed.
	- Example: OptionsView